Refactor the async wait fd logic
Implementation experience has shown that the original plan for async wait
fds was too simplistic. Originally the async logic created a pipe internally
and user/engine code could then get access to it via API calls. It is more
flexible if the engine is able to create its own fd and provide it to the
async code.

Another issue is that performance can be hit quite badly if there is a lot
of churn in the fd value within the context of (say) a single SSL
connection. It is quite expensive to be continually adding and removing fds
from (say) epoll. It is better if we can provide some stability of the fd
value across a whole SSL connection. This is problematic because an engine
has no concept of an SSL connection.

This commit refactors things to introduce an ASYNC_WAIT_CTX which acts as a
proxy for an SSL connection down at the engine layer.